fix: Clear group collector metrics by node
When the group collector collects for n nodes, we cannot clear the
prometheus metric when collecting for a particular node because this
also removes metrics for all other nodes leading to flaky metrics.

Since the public prometheus API does not provide functionality to
get or clear existing label values by a subset of values (node in our
case), we maintain the active labels by node on our own.